Revert "Merge pull request #47 from IshitaGH/main-profile-pictures"
As decided in mtg, this just reverts the changes from PR #47 so the dependies can be reorganized
When it gets remerged, it can resolve the merge conflicts in home.tsx and use the existing logic there
to get the room code from the backend.

This reverts commit 762574c30dde358f7313374ded2b9a1fe384a0b0, reversing
changes made to 9c9491734cfa1addc65e5f65f48cc43147f65492.